fix: JQPL 메서드 제거

+++ src/main/java/com/ebiz/tableorder/menu/dto/SalesDataPoint.java
@@ -8,9 +8,9 @@ import java.math.BigDecimal;
 @Getter
 public class SalesDataPoint {
     private final Integer hour;
-    private final BigDecimal revenue;
+    private final long revenue;   // ← long 으로 변경
 
-    public SalesDataPoint(Integer hour, BigDecimal revenue) {
+    public SalesDataPoint(Integer hour, long revenue) {
         this.hour    = hour;
         this.revenue = revenue;
     }
